Refactored the package to generate sonar report
