--- src/main/java/uk/gov/hmcts/reform/roleassignmentbatch1/domain/model/enums/ActorIdType.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums;
-
-public enum ActorIdType {
-    IDAM, CASEPARTY
-}
-
