--- src/main/java/uk/gov/hmcts/reform/roleassignmentbatch1/domain/model/enums/Classification.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums;
-
-public enum Classification {
-    PUBLIC, PRIVATE, RESTRICTED
-}
