+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch2/domain/model/enums/ActorIdType.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums;
+
+public enum ActorIdType {
+    IDAM, CASEPARTY
+}
+
