+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch2/domain/model/enums/RoleType.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums;
+
+public enum RoleType {
+    CASE, ORGANISATION
+}
