+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch2/domain/model/enums/Status.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums;
+package uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums;
 
 public enum Status {
     CREATED(10),
