+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch2/task/DeleteExpiredRecords.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.task;
+package uk.gov.hmcts.reform.roleassignmentbatch2.task;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
