+++ src/test/java/uk/gov/hmcts/reform/roleassignmentbatch2/helper/TestDataBuilder.java
@@ -1,16 +1,16 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.helper;
+package uk.gov.hmcts.reform.roleassignmentbatch2.helper;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Setter;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.ActorIdType;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.Classification;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.GrantType;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.RoleCategory;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.RoleType;
-import uk.gov.hmcts.reform.roleassignmentbatch1.domain.model.enums.Status;
-import uk.gov.hmcts.reform.roleassignmentbatch1.task.RoleAssignmentHistory;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.ActorIdType;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.Classification;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.GrantType;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.RoleCategory;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.RoleType;
+import uk.gov.hmcts.reform.roleassignmentbatch2.domain.model.enums.Status;
+import uk.gov.hmcts.reform.roleassignmentbatch2.task.RoleAssignmentHistory;
 
 import java.io.IOException;
 import java.io.InputStream;
